#29 split SvgDataPath generation to its own Generator
